Introduce HTTPConnector to use it from cache and FS
