+++ pfio/cache/http_cache.py
@@ -1,15 +1,8 @@
-import logging
-import os
 import pickle
-import time
-
-import urllib3
-import urllib3.exceptions
+from typing import Any
 
 from pfio.cache import Cache
-
-logger = logging.getLogger(__name__)
-logger.addHandler(logging.StreamHandler())
+from pfio.util import HTTPConnector
 
 
 class HTTPCache(Cache):
@@ -56,46 +49,8 @@ class HTTPCache(Cache):
         self.length = length
         assert self.length > 0
 
-        if url.endswith("/"):
-            self.url = url
-        else:
-            self.url = url + "/"
-
-        if bearer_token_path is not None:
-            self.bearer_token_path = bearer_token_path
-        else:
-            self.bearer_token_path = os.getenv("PFIO_HTTP_BEARER_TOKEN_PATH")
-
-        if self.bearer_token_path is not None:
-            self._token_read_now()
-
+        self.connector = HTTPConnector(url, bearer_token_path)
         self.do_pickle = do_pickle
-
-        self.conn = None
-        self._prepare_conn()
-
-        self.pid = os.getpid()
-
-    @property
-    def is_forked(self):
-        return self.pid != os.getpid()
-
-    def _checkconn(self):
-        if self.is_forked or self.conn is None:
-            self._prepare_conn()
-            self.pid = os.getpid()
-
-    def _prepare_conn(self):
-        # Allow redirect or retry once
-        self.conn = urllib3.poolmanager.PoolManager(retries=1, timeout=3)
-
-    def __getstate__(self):
-        state = self.__dict__.copy()
-        state['conn'] = None
-        return state
-
-    def __setstate__(self, state):
-        self.__dict__ = state
 
     def __len__(self):
         return self.length
@@ -108,68 +63,23 @@ class HTTPCache(Cache):
     def multithread_safe(self):
         return True
 
-    def put(self, i, data):
-        self._checkconn()
+    def put(self, i: int, data: Any):
         if i < 0 or self.length <= i:
             raise IndexError("index {} out of range ([0, {}])"
                              .format(i, self.length - 1))
         if self.do_pickle:
             data = pickle.dumps(data)
 
-        try:
-            res = self.conn.urlopen("PUT",
-                                    url=self._url(i),
-                                    headers=self._header_with_token(),
-                                    body=data)
-        except urllib3.exceptions.RequestError as e:
-            logger.warning("put: {}".format(e))
-            return False
+        return self.connector.put(str(i), data)
 
-        if res.status == 201:
-            return True
-
-        logger.warning("put: unexpected status code {}".format(res.status))
-        return False
-
-    def get(self, i):
-        self._checkconn()
+    def get(self, i: int) -> Any:
         if i < 0 or self.length <= i:
             raise IndexError("index {} out of range ([0, {}])"
                              .format(i, self.length - 1))
 
-        try:
-            res = self.conn.urlopen("GET",
-                                    url=self._url(i),
-                                    headers=self._header_with_token())
-        except urllib3.exceptions.RequestError as e:
-            logger.warning("get: {}".format(e))
-            return None
+        data = self.connector.get(str(i))
 
-        if res.status == 200:
-            if self.do_pickle:
-                return pickle.loads(res.data)
-            else:
-                return res.data
-        elif res.status == 404:
-            return None
-
-        logger.warning("get: unexpected status code {}".format(res.status))
-        return None
-
-    def _url(self, i) -> str:
-        return self.url + str(i)
-
-    def _header_with_token(self) -> dict:
-        if self.bearer_token_path is None:
-            return {}
+        if self.do_pickle and data is not None:
+            return pickle.loads(data)
         else:
-            if time.time() - self.bearer_token_updated > 1:
-                self._token_read_now()
-            return {
-                "Authorization": f"Bearer {self.bearer_token}"
-            }
-
-    def _token_read_now(self):
-        with open(self.bearer_token_path, "r") as f:
-            self.bearer_token = f.read()
-            self.bearer_token_updated = time.time()
+            return data
